+++ etna/datasets/tsdataset.py
@@ -320,6 +320,17 @@ class TSDataset:
         return df
 
     @staticmethod
+    def _cast_target_to_float(df: pd.DataFrame) -> pd.DataFrame:
+        if "target" in df.columns.get_level_values("feature").unique():
+            target_dtypes = df.loc[:, pd.IndexSlice[:, "target"]].dtypes
+            not_float_target = target_dtypes[target_dtypes != np.float64].index
+            if len(not_float_target) > 0:
+                float_target = df.loc[:, not_float_target].astype(np.float64)
+                df = df.drop(columns=not_float_target)
+                df = pd.concat([df, float_target], axis=1).sort_index(axis=1)
+        return df
+
+    @staticmethod
     def _cast_index_to_datetime(df: pd.DataFrame, freq_offset: pd.DateOffset) -> pd.DataFrame:
         if pd.api.types.is_numeric_dtype(df.index):
             warnings.warn(
@@ -339,6 +350,9 @@ class TSDataset:
 
         # cast segment to str type
         cls._cast_segment_to_str(df)
+
+        # cast target columns to float64
+        df = cls._cast_target_to_float(df)
 
         # handle freq
         if freq_offset is None:
@@ -1070,6 +1084,8 @@ class TSDataset:
         2021-01-04           3           8
         2021-01-05           4           9
         """
+        if "target" in df.columns:
+            df["target"] = df["target"].astype(np.float64)
         df = df.set_index(["timestamp", "segment"])
 
         df = df.unstack(level=-1)
@@ -1108,8 +1124,7 @@ class TSDataset:
             cur_level_to_next_level_edges = df_level_columns[[cur_level_name, next_level_name]].drop_duplicates()
             cur_level_to_next_level_adjacency_list = cur_level_to_next_level_edges.groupby(cur_level_name).agg(list)
 
-            # support for pandas>=1.4, <1.5
-            level_structure.update(cur_level_to_next_level_adjacency_list.itertuples(name=None))
+            level_structure.update(cur_level_to_next_level_adjacency_list.to_records())
             cur_level_name = next_level_name
 
         hierarchical_structure = HierarchicalStructure(
@@ -1352,6 +1367,9 @@ class TSDataset:
         Columns in ``df_exog`` are not updated. If you wish to update the ``df_exog``, create the new
         instance of TSDataset.
 
+        Updating ``df`` with ``df_update`` with different corresponding column dtypes
+        could lead to unexpected behaviour in different ``pandas`` versions.
+
         Parameters
         ----------
         df_update:
@@ -1379,8 +1397,7 @@ class TSDataset:
         try:
             column_idx = self._df.columns.get_indexer(df.columns)
 
-        # some older pandas versions <1.3 throw `ValueError`
-        except (pd.errors.InvalidIndexError, ValueError):
+        except pd.errors.InvalidIndexError:
             raise ValueError("The dataset features set contains duplicates!")
 
         original_types = df.dtypes.to_dict()
@@ -1526,12 +1543,10 @@ class TSDataset:
             target_level_df = self.to_pandas(features=target_names)
 
         target_components_df = target_level_df.loc[:, pd.IndexSlice[:, self.target_components_names]]
-        if len(self.target_components_names) > 0:  # for pandas >=1.1, <1.2
-            target_level_df = target_level_df.drop(columns=list(self.target_components_names), level="feature")
+        target_level_df = target_level_df.drop(columns=list(self.target_components_names), level="feature")
 
         prediction_intervals_df = target_level_df.loc[:, pd.IndexSlice[:, self.prediction_intervals_names]]
-        if len(self.prediction_intervals_names) > 0:  # for pandas >=1.1, <1.2
-            target_level_df = target_level_df.drop(columns=list(self.prediction_intervals_names), level="feature")
+        target_level_df = target_level_df.drop(columns=list(self.prediction_intervals_names), level="feature")
 
         ts = TSDataset(
             df=target_level_df,
@@ -1602,9 +1617,8 @@ class TSDataset:
 
     def drop_target_components(self):
         """Drop target components from dataset."""
-        if len(self.target_components_names) > 0:  # for pandas >=1.1, <1.2
-            self._df.drop(columns=list(self.target_components_names), level="feature", inplace=True)
-            self._target_components_names = ()
+        self._df.drop(columns=list(self.target_components_names), level="feature", inplace=True)
+        self._target_components_names = ()
 
     def add_prediction_intervals(self, prediction_intervals_df: pd.DataFrame):
         """Add target components into dataset.
@@ -1654,9 +1668,8 @@ class TSDataset:
 
     def drop_prediction_intervals(self):
         """Drop prediction intervals from dataset."""
-        if len(self.prediction_intervals_names) > 0:  # for pandas >=1.1, <1.2
-            self._df.drop(columns=list(self.prediction_intervals_names), level="feature", inplace=True)
-            self._prediction_intervals_names = tuple()
+        self._df.drop(columns=list(self.prediction_intervals_names), level="feature", inplace=True)
+        self._prediction_intervals_names = tuple()
 
     def isnull(self) -> pd.DataFrame:
         """Return dataframe with flag that means if the correspondent element in wide representation of data is null.