+++ etna/transforms/math/binary_operator.py
@@ -79,10 +79,14 @@ class BinaryOperationTransform(ReversibleTransform):
     - If during the operation a division by zero of a positive number occurs, writes +inf to this cell of the column, if negative - -inf, if 0/0 - nan.
     - In the case of raising a negative number to a non-integer power, writes nan to this cell of the column.
 
+    Different ``out_column`` dtype and result dtype in inplace operations
+    could lead to unexpected behaviour in different ``pandas`` versions.
+
     Examples
     --------
     >>> import numpy as np
-    >>> from etna.datasets import generate_ar_df
+    >>> from etna.datasets import generate_ar_df, TSDataset
+    >>> from etna.transforms import BinaryOperationTransform
     >>> df = generate_ar_df(start_time="2020-01-01", periods=30, freq="D", n_segments=1)
     >>> df["feature"] = np.full(30, 10)
     >>> df["target"] = np.full(30, 1)
@@ -91,24 +95,24 @@ class BinaryOperationTransform(ReversibleTransform):
     segment    segment_0
     feature      feature target
     timestamp
-    2020-01-01        10      1
-    2020-01-02        10      1
-    2020-01-03        10      1
-    2020-01-04        10      1
-    2020-01-05        10      1
-    2020-01-06        10      1
+    2020-01-01        10    1.0
+    2020-01-02        10    1.0
+    2020-01-03        10    1.0
+    2020-01-04        10    1.0
+    2020-01-05        10    1.0
+    2020-01-06        10    1.0
     >>> transformer = BinaryOperationTransform(left_column="feature", right_column="target", operator="+", out_column="target")
     >>> new_ts = transformer.fit_transform(ts=ts)
     >>> new_ts["2020-01-01":"2020-01-06", "segment_0", ["feature", "target"]]
     segment    segment_0
     feature      feature target
     timestamp
-    2020-01-01        10      11
-    2020-01-02        10      11
-    2020-01-03        10      11
-    2020-01-04        10      11
-    2020-01-05        10      11
-    2020-01-06        10      11
+    2020-01-01        10   11.0
+    2020-01-02        10   11.0
+    2020-01-03        10   11.0
+    2020-01-04        10   11.0
+    2020-01-05        10   11.0
+    2020-01-06        10   11.0
     """
 
     def __init__(self, left_column: str, right_column: str, operator: str, out_column: Optional[str] = None):