+++ etna/transforms/math/sklearn.py
@@ -28,7 +28,11 @@ class TransformMode(StringEnumWithRepr):
 
 
 class SklearnTransform(ReversibleTransform):
-    """Base class for different sklearn transforms."""
+    """Base class for different sklearn transforms.
+
+    Different ``in_column`` column dtype and result dtype with ``inplace=True`` option
+    could lead to unexpected behaviour in different ``pandas`` versions.
+    """
 
     def __init__(
         self,